+++ packages/web-sdk/src/instrumentations/session/sessionManager/sessionManagerUtils.ts
@@ -1,5 +1,7 @@
 import { dateNow, faro, genShortID } from '@grafana/faro-core';
 
+import { isLocalStorageAvailable, isSessionStorageAvailable } from '../../../utils';
+
 import { isSampled } from './sampling';
 import { SESSION_EXPIRATION_TIME, SESSION_INACTIVITY_TIME } from './sessionConstants';
 import type { FaroUserSession } from './types';
@@ -60,6 +62,13 @@ export function getUserSessionUpdater({ fetchUserSession, storeUserSession }: Ge
       return;
     }
 
+    const sessionTrackingConfig = faro.config.sessionTracking;
+    const isPersistentSessions = sessionTrackingConfig?.persistent;
+
+    if ((isPersistentSessions && !isLocalStorageAvailable) || (!isPersistentSessions && !isSessionStorageAvailable)) {
+      return;
+    }
+
     const sessionFromStorage = fetchUserSession();
 
     if (isUserSessionValid(sessionFromStorage)) {
@@ -73,7 +82,7 @@ export function getUserSessionUpdater({ fetchUserSession, storeUserSession }: Ge
       storeUserSession(newSession);
 
       faro.api?.setSession(newSession.sessionMeta);
-      faro.config.sessionTracking?.onSessionChange?.(sessionFromStorage?.sessionMeta ?? null, newSession.sessionMeta!);
+      sessionTrackingConfig?.onSessionChange?.(sessionFromStorage?.sessionMeta ?? null, newSession.sessionMeta!);
     }
   };
 }